DEVO-229: Move processed files to another folder
REF DEVO-229

Sometimes there are many files that need to be
ingested/deleted/suppressed and the task fails halfway in and then
starts right back from the beginning.

This change would avoid re-ingesting or deleting/suppressing a file that
has already been processed in an earlier failed task. Thus potentially
minimizing recovery time.

This change would also have the added benefit that in the case where
recovery is not possible because of a corrupt file, then that corrupt
file would be first on the list which can potentially reduce debugging
time.