Skip moving s3 files when src and dest are equal.
We don't have a straight forward way to update file paths for processed files
in full reindex context.  This is a quick fix to avoid error were we try
to s3 mv a file to itself (i.e. source and destination are equal).

TODO: Alma SFTP dump should go to  a subfolder so that we can easily add the
same efficiency step we use in harvest context to the full reindex context.