Fix harvest dags not working.
We made an update to the marc ingesting script that requires the list of
files to be processed be passed in via a DATA environment variable.

This change was required to allow for concurrent indexing.

This commit updates the harvest dags to set DATA env variable as
required by the new script.